% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-location.R
\name{all_files}
\alias{all_files}
\title{Scans the input directory and returns csv or xlsx files in that directory}
\usage{
all_files()
}
\value{
a dataframe containing filename, path, date, hospital, and study_year fields
}
\description{
Extracting metadata from the filename where present - particularly hospital, and year number
}
