% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework-augment.R
\name{augment_data}
\alias{augment_data}
\title{Sanitise AvonCap data columns}
\usage{
augment_data(x, ...)
}
\arguments{
\item{x}{\itemize{
\item the raw data from \code{load_data()}
}}

\item{...}{
  Named arguments passed on to \code{\link[=augment_generic]{augment_generic}}\describe{
    \item{\code{df}}{a data frame}
\item{\code{...}}{unnamed parameters are a list of functions, named parameters
are passed to those functions (if they match formal arguments).}
}}
}
\value{
a tracked dataframe with
}
\description{
AvonCap data has lots of columns which are named in a difficult to remember
fashion, composed of data items that have enumerated values with no
semantics. This makes displaying them difficult and any filtering done on the
raw data inscrutable. Depending on the source of the data some different
columns may be present due to differences in the NHS and UoB data sets. The
redcap database has some options that may be checklists and some that are
radio buttons, both of these end up with mysterious names in the data.
}
\details{
This function maps the data into a tidy dataframe with consistently named
columns, and named factors where appropriate. If not present in the data the
ethnicity

files Most of the sanitisation code is held in the
\code{zzz-avoncap-mappings.R} file.
}
