% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-framework.R
\name{augment_generic}
\alias{augment_generic}
\title{Applies a set of functions to the whole dataframe}
\usage{
augment_generic(df, ...)
}
\arguments{
\item{df}{a data frame}

\item{...}{unnamed parameters are a list of functions, named parameters
are passed to those functions (if they match formal arguments).}
}
\value{
the altered df
}
\description{
This sequences, catches errors and allows parameters to be passed by name
}
\examples{
fn1 = function(df,v) {df \%>\% dplyr::filter(cut=="Fair") \%>\% dplyr::mutate(x_col = color)}
fn2 = function(df,v) {df \%>\% dplyr::filter(color==v$color$J)}
df = diamonds \%>\% augment_generic(fn1, fn2)
}
