% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stacked-bars.R
\name{binomial_proportion_points}
\alias{binomial_proportion_points}
\title{Dodged bar and whiskers proportions}
\usage{
binomial_proportion_points(data, mapping, ..., width = 0.8, size = 0.5)
}
\arguments{
\item{data}{the data}

\item{mapping}{a aes mapping with at least \code{x} and \code{fill}. If facetting then
\code{group} must contain the facet variable}

\item{...}{passed to \code{geom_bar}}

\item{width}{width of position dodge}

\item{size}{the bar size}
}
\value{
a ggplot
}
\description{
This function plots a stacked bar of proportions for an input set of data
}
