% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-labelling.R
\name{default_column_names}
\alias{default_column_names}
\title{default column naming mappings}
\usage{
default_column_names(...)
}
\arguments{
\item{...}{additional named items to add}
}
\value{
a set of mappings
}
\description{
default column naming mappings
}
