% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_admission_episode}
\alias{derive_admission_episode}
\title{Create a counter in the event of repeated admissions}
\usage{
derive_admission_episode(df, v)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}
}
\value{
a dataframe
}
\description{
This also will calculate a time interval between admissions.
There is also a repeat admission instrument that this does not use.
}
\concept{derived}
