% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_aLRTD_categories}
\alias{derive_aLRTD_categories}
\title{The aLRTD incidence paper classifications}
\usage{
derive_aLRTD_categories(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
The 3 category classifications
}
\details{
\itemize{
\item aetiological:
\itemize{
\item Confirmed SARS-CoV-2 - implies Infective
\item No evidence SARS-CoV-2 - implies Infective but not confirmed as SARS-CoV-2
\item Non-infective - presumed non infective
}
\item clinical presentation:
\itemize{
\item Pneumonia - implies Infective
\item NP-LRTI - implies Infective
\item No evidence LRTI (include CRDE and HF)
}
}

Some cases do not get a clinical presentation in this. Typically they are people
who have an infective cause, but LRTI and pneumonia have been excluded. These
could be URTI and or incidental COVID cases.
}
\concept{derived}
