% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_catchment_status}
\alias{derive_catchment_status}
\title{Identify patients who are in the BNSSG ICB based on their GP practice name}
\usage{
derive_catchment_status(df, v)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}
}
\value{
a dataframe
}
\description{
Names are normalised by removing commonly mixed up components and
}
\concept{derived}
