% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_completed_vaccination_status}
\alias{derive_completed_vaccination_status}
\title{Derive detailed vaccination status on admission}
\usage{
derive_completed_vaccination_status(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
Vaccination is deemed to have had effect if given > 14 days before admission
for 1st dose or >7 days before admission for subsequent doses.
This does not account for previous infection which is not in the data set.
}
\concept{derived}
