% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_pneumococcal.R
\name{derive_continuous_categories_pneumo}
\alias{derive_continuous_categories_pneumo}
\title{Age and CURB score categories}
\usage{
derive_continuous_categories_pneumo(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
This should be consistent with AvonCAP age / CURB cateories.
}
\concept{derived}
