% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_diagnosis_categories}
\alias{derive_diagnosis_categories}
\title{Create 4 non exclusive diagnostic categories}
\usage{
derive_diagnosis_categories(df, v)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}
}
\value{
a dataframe
}
\description{
Pneumonia if one of:
\itemize{
\item Standard of care diagnosis of CAP (radiologically or clinically)
\item Empyema or abscess
\item Admission chest X-ray shows pneumonia
}
}
\details{
NP-LRTI if:
\itemize{
\item Not pneumonia and Standard of care LTRI diagnosis
}

Exascerbation of CRDE:
\itemize{
\item Standard of care exacerbation COPD
\item Standard of care exacerbation Non-COPD
\item (N.B. may be pneumonia or NP-LRTI)
}

Heart failure:
\itemize{
\item Standard of care congestive heart failure.
}
}
\concept{derived}
