% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_effective_vaccination_status}
\alias{derive_effective_vaccination_status}
\title{A simple vaccination status on admission as an ordered number of doses}
\usage{
derive_effective_vaccination_status(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
This does not account for previous infection which is not in the data set.
}
\concept{derived}
