% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_genomic_variant}
\alias{derive_genomic_variant}
\title{Give a inferred Alpha, Delta or Omicron status based on time alone.}
\usage{
derive_genomic_variant(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
This relies on date period during which we are very confidence that the only
variants circulating are of a given type. These are quite conservative estimates
based on the frequency of sequenced cases in the bristol area (according to the
Sanger centre and to cases identified in the hospital testing)
}
\details{
\href{https://covid19.sanger.ac.uk/lineages/raw?date=2021-07-24&area=E06000023&lineageView=1&lineages=A\%2CB\%2CB.1.1.7\%2CB.1.617.2\%2CB.1.1.529&colours=7\%2C3\%2C1\%2C6\%2C2}{Sanger centre data}
\itemize{
\item Pre-alpha before 05 Dec 2020
\item Alpha between 13 Feb 2021 and 15 May 2021
\item Delta between 01 Jun 2021 and 07 Nov 2021
\item Omicron from 07 Feb 2022 to present
}
}
\concept{derived}
