% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_gp_linkage}
\alias{derive_gp_linkage}
\title{Identify patients from surgeries in linked GP study}
\usage{
derive_gp_linkage(df, v)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}
}
\value{
a dataframe
}
\description{
Identify patients from surgeries in linked GP study
}
\concept{derived}
