% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_haematology_categories}
\alias{derive_haematology_categories}
\title{Binary outcomes for haematology data}
\usage{
derive_haematology_categories(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
\itemize{
\item Evelated troponin : > 14
}
}
\concept{derived}
