% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_hospital_burden_outcomes}
\alias{derive_hospital_burden_outcomes}
\title{Binary outcomes for hospital burden}
\usage{
derive_hospital_burden_outcomes(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
These outcomes were tested in the Delta vs Omicron severity paper and
sensitiivyt analysis
}
\details{
\itemize{
\item O2 requirement within 7 days (various cut-offs)
\item Any respiratory support in 7 days (various cut-offs)
\item LOS > X days in first 7 days (various cut-offs)
}
}
\concept{derived}
