% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_infective_classification}
\alias{derive_infective_classification}
\title{Determine if an admission is due to an infective cause}
\usage{
derive_infective_classification(df, v)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}
}
\value{
a dataframe
}
\description{
Infective admissions are defined as any of:
\itemize{
\item pneumonias
\item NP-LRTI
\item laboratory confirmed COVID diagnosis
\item admission swab COVID positive
}
}
\details{
Infective admissions are excluded if:
\itemize{
\item Standard of care states non-infectious process
\item SOC non-LRTI (and none of the other categories above)
}

Any unknowns are defined as non-Infective
}
\concept{derived}
