% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_pneumococcal.R
\name{derive_invasive_status}
\alias{derive_invasive_status}
\title{Pneumococcal invasive status and binary test category}
\usage{
derive_invasive_status(df, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
Pneumococcal invasive status and binary test category
}
\concept{derived}
