% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_nosocomial_covid_status}
\alias{derive_nosocomial_covid_status}
\title{Did the patient catch COVID in hospital}
\usage{
derive_nosocomial_covid_status(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
Only relevant to SARS-CoV-2 PCR positive patient.
Timing of positive test compared to admission:
This relies on knowing dates and hence only works on the identifiable data
sets,
}
\details{
Logic is:
\itemize{
\item Community if PCR result predates admission
\item Probably commuinity if PCR result within 7 days of admission
\item Probably nosocomial if 7-28 days after admission
\item Otherwise is it undefined.
}
}
\concept{derived}
