% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_nosocomial_status}
\alias{derive_nosocomial_status}
\title{Identify patients who were admitted already prior to study entry}
\usage{
derive_nosocomial_status(df, v)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}
}
\value{
a dataframe
}
\description{
Hospital acquired COVID is recorded explicitly in 2 places for some patients.
A large difference between admission date and enrollment date (<21 days) is
suggestive in other cases
}
\concept{derived}
