% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-generic.R
\name{derive_pandemic_timings}
\alias{derive_pandemic_timings}
\title{Date columns}
\usage{
derive_pandemic_timings(date_col, prefix)
}
\arguments{
\item{date_col}{the date column}

\item{prefix}{a prefix for the columns to be added}
}
\value{
a \code{derive_...} style function to augment a data set containing \code{date_col}
with a set of columns describing the timing.
}
\description{
Date columns
}
