% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-pneumococcal.R
\name{derive_pcv_groupings}
\alias{derive_pcv_groupings}
\title{Group pneumo serotypes according to e.g. vaccine coverage}
\usage{
derive_pcv_groupings(
  df,
  ...,
  pcv_map = default_pcv_map,
  not_matched = "Other",
  col_name = "pneumo.pcv_group"
)
}
\arguments{
\item{df}{the normalised urine antigen data}

\item{...}{ignored}

\item{pcv_map}{a 2 column data frame mapping \code{group} to \code{serotype}}

\item{not_matched}{what to call the column of non-matched serotypes? Default is
\code{Other}, but \verb{Non vaccine type} might be preferred.}

\item{col_name}{the target column name for the pcv grouping (defaults
to \code{pneumo.pcv_group})}
}
\value{
an augmented data frame with an additional column defined by \code{col_name}
}
\description{
A range of useful serotype groups is defined in the list \code{uad_groups}. The
\code{default_pcv_map} gives a set of mappings to group headings that gives the
overall serotype distribution by vaccine.
}
\details{
The logic employed in combining elements is:
\itemize{
\item any(result == "Unknown") ~ "Unknown"
\item any(result == "Positive") ~ "Positive"
\item all(result == "Negative") ~ "Negative"
\item TRUE ~ "Other"
}
}
\concept{derived}
