% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_pneumococcal.R
\name{derive_phe_pcv_group}
\alias{derive_phe_pcv_group}
\title{Get vaccine coverage group for known serotype}
\usage{
derive_phe_pcv_group(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
For the longitudinal oneumocococcal data, a range of useful serotype groups
is defined in the list \code{avoncap::serotype_data}. The \code{avoncap::serotype_pcv_map} gives a set of
mappings to (multiple) group headings that gives the overall serotype distribution by
vaccine.
}
\concept{derived}
