% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_pneumococcal.R
\name{derive_pneumo_clinical_syndrome}
\alias{derive_pneumo_clinical_syndrome}
\title{Add in clinical syndrome indicator}
\usage{
derive_pneumo_clinical_syndrome(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
A list of presentations based on site which
\itemize{
\item LRTI
\item Meningitis
\item Effusion/Empyema
\item Septic arthritis
\item URTI
\item Other
}
}
\concept{derived}
