% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-pneumococcal.R
\name{derive_pneumo_polyfill}
\alias{derive_pneumo_polyfill}
\title{Make pneumo data compatible with AvonCAP}
\usage{
derive_pneumo_polyfill(df, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
Needed for:
\itemize{
\item \code{derive_simpler_comorbidities}
\item \code{derive_pneumococcal_high_risk}
\item \code{derive_pneumococcal_risk_category}
}
}
\concept{derived}
