% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-uad.R
\name{derive_pneumo_uad_panel}
\alias{derive_pneumo_uad_panel}
\title{Calculate UAD panel for test}
\usage{
derive_pneumo_uad_panel(df, ...)
}
\arguments{
\item{df}{a pneumo serotype dataframe}

\item{...}{ignored}
}
\value{
a dataframe with additional columns \code{pneumo.uad1_panel_result},
\code{pneumo.uad2_panel_result}, \code{pneumo.non_uad_panel_result},
\code{pneumo.serotype_summary_result}
}
\description{
The panels are UAD1 for PCV13 serotypes, UAD2 for PPV23 serotypes.
}
\concept{derived}
