% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_pneumococcal_categories}
\alias{derive_pneumococcal_categories}
\title{The pneumococcal incidence diagnostic classifications}
\usage{
derive_pneumococcal_categories(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
The 4 category disjoint classification.
}
\details{
\itemize{
\item pneumo.presentation_class:
\itemize{
\item CAP+/RAD+ - radiologically proved pneumonia
\item CAP+/RAD- - pneumonia without x-ray confirmation
\item NP-LRTI - non-pneumonic lower respiratory tract infection
\item No evidence LRTI - believed to be non-infective at admission, this last
group is usually discarded from analysis, however it only really describes
people without a clinical diagnosis of infection on admission. There could
still be undiagnosed infection there.
}
}
}
\concept{derived}
