% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_pneumococcal_high_risk}
\alias{derive_pneumococcal_high_risk}
\title{Determine if patient is in a high pneumococcal risk group}
\usage{
derive_pneumococcal_high_risk(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
High pneumococcal risk defined if any of the following:
\itemize{
\item over 65 years old
\item other pneumococcal risks
\item comorbid copd
\item interstitial lung disease
\item cystic fibrosis
\item hypertension
\item CCF
\item ischaemic heart disease
\item chronic kidney disease
\item chronic liver disease
\item diabetes
\item asthmatic with immunodeficiency
\item on immunosupression
}
}
\concept{derived}
