% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_polyfill_central}
\alias{derive_polyfill_central}
\title{Polyfill data}
\usage{
derive_polyfill_central(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
Some basic context to allow comparison to ED data.
}
\details{
\itemize{
\item All of the patients admitted
}
}
\concept{derived}
