% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-ed.R
\name{derive_polyfill_ed}
\alias{derive_polyfill_ed}
\title{Polyfill ED data}
\usage{
derive_polyfill_ed(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
The ED data has some different fields from the main avoncap data.
}
\details{
\itemize{
\item It is missing an admission cxr summary field needed to calculate pneumonia
\item It has a fixed admission route of "A&E" (i.e. ED to non UK people)
\item None of the patients admitted
\item Hospital admission length of stay is zero
}
}
\concept{derived}
