% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_presumed_diagnosis_categories}
\alias{derive_presumed_diagnosis_categories}
\title{Create presumed diagnostic categories}
\usage{
derive_presumed_diagnosis_categories(df, v)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}
}
\value{
a dataframe
}
\description{
Pneumonia if one of:
\itemize{
\item Initial diagnosis of CAP (supported by initial radiology or clinically)
\item Empyema or abscess
}
}
\details{
Presumed clinical presentation:
\itemize{
\item Pneumonia - implies Infective
\item NP-LRTI - implies Infective
\item No evidence LRTI (include CRDE and HF)
}
}
\concept{derived}
