% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-qcovid2.R
\name{derive_qcovid}
\alias{derive_qcovid}
\title{Calculate a QCOVID2 score from AvonCap data source}
\usage{
derive_qcovid(df, v = avoncap_df \%>\% get_value_sets())
}
\arguments{
\item{avoncap_df}{a normalised augmented avoncap data source}
}
\value{
the same dataframe with additional columns,
\itemize{
\item qcovid2.log_hazard, covid2.hazard_ratio: a log hazard rate for the QCOVID2
score where missing data is substituted with the reference value for the QCOVID2
population.
\item qcovid2.log_comorbid_hazard, qcovid2.comorbid_hazard_ratio: a log hazard
rate for the comorbid conditions and not including age and BMI.
}
}
\description{
uses inbuilt imd_to_townsend map. This implements a cut down version of the
QCovid2 score depending on what data is available.
}
