% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_generic.R
\name{derive_quintile_category}
\alias{derive_quintile_category}
\title{Split a continuous variable into quintiles}
\usage{
derive_quintile_category(col, labels = c("1-short", "2", "3", "4", "5-long"))
}
\arguments{
\item{col}{the continuous data column that is to be categorised by quintile.}

\item{labels}{the category labels}
}
\value{
a \code{derive_...} style function that augments a data set with col \code{xxx}
with col \code{xxx_quintile} containing the quintiles
}
\description{
Split a continuous variable into quintiles
}
