% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_simpler_comorbidities}
\alias{derive_simpler_comorbidities}
\title{Rationalise some of the more detailed comorbidities}
\usage{
derive_simpler_comorbidities(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
Rationalise some of the more detailed comorbidities
}
\concept{derived}
