% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_survival_censoring}
\alias{derive_survival_censoring}
\title{Survival outcomes}
\usage{
derive_survival_censoring(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
Calculates:
}
\details{
\itemize{
\item A consistent length of stay - shortest of length of stay and 30 day and 1 yr survival duration
\item a 30 day survival duration and censoring status for survfit
\item a 1 year survival duration and censoring status for survfit
\item Hospital length of stay and censoring status for survfit
\item Categorical length of stay and 30 day survival 0-3, 4-6, 7-13, 14-29, gte 30
}

time: for this is the follow up time to event.

event: The event type indicator
\itemize{
\item 0=alive/censored,
\item 1=dead.
}

or for length of stay
\itemize{
\item 0=still inpatient,
\item 1=discharged from hospital/ dead
}
}
\concept{derived}
