% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-pneumococcal.R
\name{derive_survival_times_pneumo}
\alias{derive_survival_times_pneumo}
\title{Survival analysis times}
\usage{
derive_survival_times_pneumo(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
Fixes a data issue with length of stay and survival duration being filled in
across 2 columns. and missing last observation dates so that we can calculate
survival censoring consistently in other data sets.
}
\details{
Calculates:
\itemize{
\item A consistent length of stay - shortest of length of stay and 30 day and 1 yr survival duration
\item A consistent uncensored time to death - shortest of 30 day and 1 yr survival duration
\item A consistent time to last observation
}
}
\concept{derived}
