% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework-augment.R
\name{derive_template}
\alias{derive_template}
\title{Derived data function template}
\usage{
derive_template(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
Derived data function template
}
