% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_vaccination_timings}
\alias{derive_vaccination_timings}
\title{Derive times from vaccination to symptom onset}
\usage{
derive_vaccination_timings(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
If symptom duration is not given it is assumed to be zero.
}
\concept{derived}
