% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_vaccine_combinations}
\alias{derive_vaccine_combinations}
\title{Deprecated - Vaccine combinations are less relevant now}
\usage{
derive_vaccine_combinations(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
There are too many potential combinations with 4th, 5th and sixth dose to make
this useful.
}
\concept{derived}
