% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_WHO_outcome_score}
\alias{derive_WHO_outcome_score}
\title{determine WHO outcome score}
\usage{
derive_WHO_outcome_score(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
Scores 0-3 are for community cases.
}
\details{
We generally cant tell the difference between 7 and 8.
\itemize{
\item 4: Hospitalised; no oxygen therapy
\item 5: Hospitalised; oxygen by mask or nasal prongs
\item 6: Hospitalised; oxygen by NIV or high flow
\item 7: Intubation and mechanical ventilation, pO2/FiO2 ≥150 or SpO2/FiO2 ≥200
\item 8: Mechanical ventilation pO2/FIO2 <150 (SpO2/FiO2 <200) or vasopressors
\item 9: Mechanical ventilation pO2/FiO2 <150 and vasopressors, dialysis, or ECMO
\item 10: Dead
}
}
\concept{derived}
