% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-labelling.R
\name{extract_dependencies}
\alias{extract_dependencies}
\title{Get provenance of data column}
\usage{
extract_dependencies(data, col, original = TRUE)
}
\arguments{
\item{data}{the dataframe}

\item{col}{the column as a symbol}

\item{original}{map the names to the original column names from the data. If
this is false the function returns a list of current normalised column names.}
}
\value{
a named list of dependencies and original column names for a given
column
}
\description{
When a data set is normalised or augmented the original column names are
stored as metadata. This helps us determine how a particular item was created.
In future this will be useful for documentation.
}
