% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-labelling.R
\name{find_new_field_names}
\alias{find_new_field_names}
\title{Get the transformed columns from original field names}
\usage{
find_new_field_names(normalised, fields)
}
\arguments{
\item{normalised}{the transformed data set.}

\item{fields}{a vector of field names}
}
\value{
a named list mapping original to new columns
}
\description{
Get the transformed columns from original field names
}
