% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{frameworks}
\alias{frameworks}
\title{Frameworks}
\description{
The list of validation, normalisation and augmentation frameworks.
There should be one validation per data set. The may be mulitple
normalisations and augmentations depending on the aspect of the data
we are extracting (e.g. re-nesting flattened data.)
}
