% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icb_surgeries}
\alias{icb_surgeries}
\title{GP surgeries in the Bristol ICB area}
\format{
A dataframe containing the following columns:
\itemize{
\item code - an official ODS code for the GP surgery
\item name - the official surgery name.
}

82 rows and 2 columns
}
\usage{
data(icb_surgeries)
}
\description{
The denominator relates only to patients coming from these GP surgeries
}
\keyword{datasets}
