% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-location.R
\name{input}
\alias{input}
\title{Locate the input directory}
\usage{
input(...)
}
\arguments{
\item{...}{the sub paths within the input directory}
}
\value{
a path to the input directory and sub paths if provided
}
\description{
Locate the input directory
}
\examples{

# devtools::load_all()

avoncap::set_input("~/Data/avoncap")
avoncap::input("nhs-extract")

avoncap::all_files()


# exact match on filename column of all_data()
avoncap::most_recent_files("AvonCAPLRTDCentralDa")


# or matches by lower case startWith on directory
avoncap::most_recent_files("nhs-extract","deltave")


avoncap::most_recent_files("metadata")

avoncap::valid_inputs()
}
