% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{key_dates}
\alias{key_dates}
\title{Key dates:}
\description{
A list of key dates:
\itemize{
\item mortality_updated - the last time the NHS mortality data was extracted and added to AvonCAP
\item min_alpha - earliest observation of the alpha variant
\item max_wuhan - last observation of the wuhan variant
\item min_delta - earliest observation of the delta variant
\item max_alpha - last observation of the alpha variant
\item min_omicron - earliest observation of the omicron variant
\item max_delta - last observation of the delta variant
}
}
