% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-km-plot.R
\name{km_plot}
\alias{km_plot}
\title{Faceted Kaplan-Meier plot}
\usage{
km_plot(
  df,
  coxmodel,
  facet = NULL,
  ...,
  maxtime = NULL,
  ylab = if (!invert) "surviving (\%)" else "affected (\%)",
  xlab = "time (days)",
  facetlab = NULL,
  ylim = (if (invert) c(0, NA) else c(NA, 100)),
  n_breaks = 5,
  heights = c(10, 1),
  invert = FALSE,
  show_label = FALSE,
  show_legend = TRUE
)
}
\arguments{
\item{df}{the data}

\item{coxmodel}{the cox model output of survival::coxph from the data}

\item{facet}{the division to highlight in the KM strata. Defaults to first
term on the lhs of the cox model formula}

\item{...}{
  Named arguments passed on to \code{\link[survival:survfit]{survival::survfit}}\describe{
    \item{\code{formula}}{either a formula or a previously fitted model}
\item{\code{...}}{other arguments to the specific method}
}}

\item{maxtime}{the longest x value to plot (optional)}

\item{ylab}{the y axis label}

\item{xlab}{the x axis label}

\item{facetlab}{a label to add as a facet title}

\item{ylim}{the range to show on the KM plot}

\item{n_breaks}{number of x axis breaks to display this also determines the
timing and number of "at risk" counts to display.}

\item{heights}{the relative height between the KM plot and the "at risk" table}

\item{invert}{reverse survival statistics to count number of affected}

\item{show_label}{show the label on the at risk table ( which is somewhat
redundant as items are coloured)}

\item{show_legend}{show the legend for the strata. (This is sometimes redundant
if the at risk table is labelled)}
}
\value{
a ggplot patchwork.
}
\description{
Faceted Kaplan-Meier plot
}
\examples{

cox = survival::coxph(survival::Surv(time, status) ~ trt + celltype + karno +
  diagtime + age + prior , data = survival::veteran)

km_plot(survival::veteran, cox)
km_plot(survival::veteran, cox, facet = 1)

km_plot(survival::veteran, cox, "celltype", show_label=TRUE) &
   ggplot2::theme(legend.position="bottom")

km_plot(survival::veteran, cox, "trt", show_label=TRUE) &
   ggplot2::theme(legend.position="bottom")

}
