% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-location.R
\name{load_data}
\alias{load_data}
\title{Load data and check structure}
\usage{
load_data(
  type,
  subtype = NULL,
  reproduce_at = as.Date(getOption("reproduce.at", default = Sys.Date())),
  merge = TRUE,
  ...
)
}
\arguments{
\item{type}{the file category see \code{valid_inputs()} for current list in input
directory}

\item{subtype}{the subtype from \code{valid_inputs()}}

\item{reproduce_at}{\itemize{
\item the date at which to cut off newer data files
}}

\item{merge}{\itemize{
\item setting to \code{TRUE} forces multiple files be merged into a
single data frame by losing mismatching columns.
}}

\item{...}{\itemize{
\item passed to \code{cached} may specifically want to use `.nocache=TRUE``
}}
}
\value{
either a list of dataframes or a single merged dataframe
}
\description{
Loads the AvonCap data from a set of csv files, which may optionally be
qualified by site \verb{('BRI' or 'NBT')} and database year \verb{('y1', 'y2', 'y3')}
as part of the file name. This selects the most recent files earlier than the
\code{reproduce_at} date and detects whether they are in a set of files.
}
\details{
The files are loaded as csv as checked that files have (A) the same columns,
(B) the same type (or are empty) (C) have any major parse issues. It then
merges the files into a single dataframe, if possible, otherwise it will
return the individually loaded files as a list of dataframes.
}
\examples{
try(load_data("nhs-extract","deltave"))
}
