% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping-avoncap-ed.R
\name{map_avoncap_ed}
\alias{map_avoncap_ed}
\title{Avoncap ED normalisation}
\usage{
map_avoncap_ed()
}
\value{
a list
}
\description{
All the ED data is also mapped using the \code{map_avoncap_central()} list
as it si quite similar
}
\details{
\itemize{
\item ed_hours -> outcome.emergency_dept_length_of_stay (name)
\item ed_reattendance -> admin.ed_episodes_in_last_30_days (name)
\item hosp_adm_30d -> outcome.admitted_within_30_days (yesno)
\item hosp_adm_7d -> outcome.admitted_within_7_days (yesno)
\item home_d_1 -> outcome.days_since_last_ed_episode (name)
\item radiology_result_1___2 -> radio.consistent_with_pneumonia_1 (yesno)
\item radiology_result_2___2 -> radio.consistent_with_pneumonia_2 (yesno)
}
}
\concept{map}
