% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping-avoncap-ed.R
\name{map_avoncap_ed_consent}
\alias{map_avoncap_ed_consent}
\title{ED consent}
\usage{
map_avoncap_ed_consent()
}
\value{
a list
}
\description{
\itemize{
\item consented -> admin.consented (list)
\item ppc -> admin.pp_consented (list)
\item withdrawal -> admin.withdrawal (yesno)
\item consent_urine -> admin.consent_for_urine (yesno)
\item consent_blood -> admin.consent_for_blood (yesno)
\item consent_resp_samples1 -> admin.consent_for_respiratory_samples (yesno)
}
}
\concept{map}
