% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping-avoncap-haematology.R
\name{map_avoncap_haem}
\alias{map_avoncap_haem}
\title{Normalise the avoncap data haematology data}
\usage{
map_avoncap_haem()
}
\value{
a list
}
\description{
\itemize{
\item record_number -> admin.record_number (name)
\item ac_study_number -> admin.consented_record_number (study_id)
\item ph_7_35 -> haem.blood_gas_ph (double)
\item glucose -> haem.glucose (double)
\item albumin -> haem.albumin (double)
\item wcc -> haem.white_cell_count (double)
\item eos -> haem.eosinophils (double)
\item hb -> haem.haemoglobin (double)
\item haematocrit -> haem.haemotocrit (double)
\item pmn -> haem.neutrophils (double)
\item lymphocytes -> haem.lymphocytes (double)
\item crp -> haem.crp (double)
\item na_result -> haem.sodium (double)
\item ur_result -> haem.urea (double)
\item egfr -> haem.egfr (double)
\item sars_cov2_antigen -> haem.sars_cov2_antigen (trunc_double)
\item ferritin -> haem.ferritin (double)
\item troponin -> haem.troponin (double)
\item nt_probnp -> haem.pro_bnp (double)
\item d_dimer -> haem.d_dimer (double)
\item patient_blood_group -> haem.blood_group (list)
}
}
\concept{map}
