% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-avoncap-radio-mappings.R
\name{map_avoncap_radio}
\alias{map_avoncap_radio}
\title{Normalise the avoncap data radiology data}
\usage{
map_avoncap_radio(instrument)
}
\value{
a list
}
\description{
\itemize{
\item radio_exam -> radio.test_performed (yesno)
\item radiology_date -> radio.test_date (date)
\item radiodays -> radio.test_days_from_admission (name)
\item radio_test -> radio.test_type (list)
\item radiology_result -> radio.alrtd_finding (checkboxes_to_nested_list)
\item radiology_other_result -> radio.non_alrtd_finding (checkboxes_to_nested_list)
\item record_number -> admin.record_number (name)
\item ac_study_number -> admin.consented_record_number (study_id)
}
}
\concept{map}
