% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework-normalise.R
\name{normalise_data}
\alias{normalise_data}
\title{Sanitise AvonCap data columns}
\usage{
normalise_data(rawData, instrument = NULL, ...)
}
\arguments{
\item{rawData}{\itemize{
\item the raw data from \code{load_data()}
}}

\item{instrument}{the numeric instrument number if applicable}

\item{...}{
  Named arguments passed on to \code{\link[=normalise_generic]{normalise_generic}}\describe{
    \item{\code{remove_mapped}}{gets rid of original columns for which we have a mapping
(leaving the new versions)}
\item{\code{remove_unmapped}}{gets rid of columns for which we do not have a mapping}
\item{\code{mappings}}{a set of mappings (see \code{zzz-avoncap-mappings.R})}
\item{\code{messages}}{a set of \code{dtrackr} glue specs that populate the first box fo
the flow chart. (can use \code{{files}}, \code{{reproduce_at}}, \code{{date}}, \code{{.total}})}
\item{\code{data_source_info}}{\itemize{
\item if not null a filename, and the function will write
out a file with the details of the input files used.
}}
\item{\code{...}}{passed onto \code{.cached(...)}. e.g. \code{nocache = TRUE} can be used to
defeat caching.}
}}
}
\value{
a tracked dataframe with n
}
\description{
AvonCap data has lots of columns which are named in a difficult to remember
fashion, composed of data items that have enumerated values with no
semantics. This makes displaying them difficult and any filtering done on the
raw data inscrutable. Depending on the source of the data some different
columns may be present due to differences in the NHS and UoB data sets. The
redcap database has some options that may be checklists and some that are
radio buttons, both of these end up with mysterious names in the data.
}
\details{
This function maps the data into a tidy dataframe with consistently named
columns, and named factors where appropriate. The mapping is defined in data.

files Most of the sanitisation code is held in the
\code{normalise-xxx.R} file. but these in turn may depend on the \code{mapping-xxx.R}
files
}
