% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise-framework.R
\name{normalise_generic}
\alias{normalise_generic}
\title{Sanitise RedCap data columns}
\usage{
normalise_generic(
  rawData,
  mappings,
  remove_mapped = TRUE,
  remove_unmapped = TRUE,
  messages = c("files: {files}", "{.total} rows", "files from: {date}"),
  data_source_info = NULL,
  ...
)
}
\arguments{
\item{rawData}{\itemize{
\item the raw data from \code{load_data()}
}}

\item{mappings}{a set of mappings (see \code{zzz-avoncap-mappings.R})}

\item{remove_mapped}{gets rid of original columns for which we have a mapping
(leaving the new versions)}

\item{remove_unmapped}{gets rid of columns for which we do not have a mapping}

\item{messages}{a set of \code{dtrackr} glue specs that populate the first box fo
the flow chart. (can use \code{{files}}, \code{{reproduce_at}}, \code{{date}}, \code{{.total}})}

\item{data_source_info}{\itemize{
\item if not null a filename, and the function will write
out a file with the details of the input files used.
}}

\item{...}{passed onto \code{.cached(...)}. e.g. \code{nocache = TRUE} can be used to
defeat caching.}
}
\value{
a tracked dataframe with
}
\description{
redcap database has some options that may be checklists and some that are
radio buttons, both of these end up with mysterious names in the data.
}
\details{
This function maps the data into a tidy dataframe with consistently named
columns, and named factors where appropriate. If not present in the data the
ethnicity
}
