% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-labelling.R
\name{original_field_names}
\alias{original_field_names}
\title{Get the mapping of transformed columns back to original}
\usage{
original_field_names(data, inverse = TRUE)
}
\arguments{
\item{data}{the transformed data set.}

\item{inverse}{give the data as a old -> new mapping for finding
normalised names of original columns. if false gives it as new->old for finding
original names of normalised columns}
}
\value{
a named list mapping original to new columns
}
\description{
Get the mapping of transformed columns back to original
}
