% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-labelling.R
\name{readable_label}
\alias{readable_label}
\title{Get a label for a column}
\usage{
readable_label(columnVar, colNames = default_column_names())
}
\arguments{
\item{columnVar}{the column name as a string}

\item{colNames}{bespoke column names mapping (see \code{default_column_names(...)})}
}
\value{
a mapped column name
}
\description{
Get a label for a column
}
