% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wrangling.R
\name{readable_label_mapping}
\alias{readable_label_mapping}
\alias{readable_label_mapping.data.frame}
\alias{readable_label_mapping.list}
\alias{readable_label_mapping.character}
\alias{readable_label_mapping.default}
\title{Get a readable label for the AvonCap data as a named list (for ggplot)}
\usage{
readable_label_mapping(x, ...)

\method{readable_label_mapping}{data.frame}(x, colNames = default_column_names(...), ...)

\method{readable_label_mapping}{list}(x, colNames = default_column_names(...), ...)

\method{readable_label_mapping}{character}(x, colNames = default_column_names(...), ...)

\method{readable_label_mapping}{default}(x, colNames = default_column_names(...), ...)
}
\arguments{
\item{x}{either the column names as strings, or a dataframe}

\item{...}{ignored}

\item{colNames}{a mapping to convert a column name (as a string) to a readable label}
}
\value{
a named list of the labels for the columns
}
\description{
Get a readable label for the AvonCap data as a named list (for ggplot)
}
\section{Methods (by class)}{
\itemize{
\item \code{readable_label_mapping(data.frame)}: for data frames

\item \code{readable_label_mapping(list)}: for lists

\item \code{readable_label_mapping(character)}: for character vectors

\item \code{readable_label_mapping(default)}: defaults

}}
