% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-serotype-scales.R
\name{relevel_serotypes}
\alias{relevel_serotypes}
\title{Relevel serotype data into an factor based on PCV group status
and serotype name.}
\usage{
relevel_serotypes(serotypes, ..., exprs)
}
\arguments{
\item{serotypes}{a vector of serotypes as a factor or character.}

\item{...}{an unwrapped version of the \code{exprs} parameter}

\item{exprs}{a list of formulae with a predicate on the LHS and a PCV group name on
the RHS. which are interpreted as the parameters for a \code{dplyr::case_when}
call. This must be protected against interpretation by wrapping it in
\code{rlang::exprs()}. The predicates are tested against \code{avoncap::serotype_data$map}
and could use any of  the following columns 'serotype', 'PCV7', 'PCV13', 'PCV15', 'PCV20', 'PPV23', 'PCV13on7', 'PCV15on13', 'PCV20on15', 'PPV23on20', 'PCV10SSI', 'PCV10GSK', 'PCV15Zhifei', 'PCV24Vaxcyte', 'PCV24Affinivax'
a default option of the form \code{TRUE ~ "Non PCV serotype"} must exist to capture
unmatched items.}
}
\description{
Relevel serotype data into an factor based on PCV group status
and serotype name.
}
\examples{
x = rlang::exprs(
  PCV7 ~ "PCV7",
  PCV15 ~ "PCV15-7",
  TRUE ~ "Non-PCV15 serotype"
)
relevel_serotypes(avoncap::phe_serotypes, exprs=x)
relevel_serotypes(avoncap::phe_serotypes)

relevel_serotypes(avoncap::phe_serotypes,
  PCV24Affinivax ~ "Affinivax",
  TRUE ~ "Non-affinivax"
)
}
