% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-location.R
\name{save_data_source_info}
\alias{save_data_source_info}
\title{Write file source information out to a text files}
\usage{
save_data_source_info(..., .file)
}
\arguments{
\item{...}{A list of data frames loaded with the \code{load_data(...)} call}

\item{.file}{the output file location}
}
\value{
the file name written (invisibly)
}
\description{
Write file source information out to a text files
}
