% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-serotype-scales.R
\name{scale_fill_serotype}
\alias{scale_fill_serotype}
\title{A ggplot scale for pneumococcal serotypes that keeps PCV groups together}
\usage{
scale_fill_serotype(
  ...,
  palette_fn = scales::brewer_pal(palette = "Dark2"),
  undefined = "#606060",
  exprs = rlang::exprs()
)
}
\arguments{
\item{...}{an unwrapped version of the \code{exprs} parameter
  Named arguments passed on to \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual}}\describe{
    \item{\code{values}}{a set of aesthetic values to map data values to. The values
will be matched in order (usually alphabetical) with the limits of the
scale, or with \code{breaks} if provided. If this is a named vector, then the
values will be matched based on the names instead. Data values that don't
match will be given \code{na.value}.}
\item{\code{aesthetics}}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful, for
example, to apply colour settings to the \code{colour} and \code{fill} aesthetics at the
same time, via \code{aesthetics = c("colour", "fill")}.}
\item{\code{breaks}}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks (the scale limits)
\item A character vector of breaks
\item A function that takes the limits as input and returns breaks
as output
}}
\item{\code{na.value}}{The aesthetic value to use for missing (\code{NA}) values}
}}

\item{palette_fn}{a function that returns a set of colours for a number of
levels. Such functions can be obtained from things like \code{scales::brewer_pal(...)}}

\item{undefined}{the colour for the last group which is assumed to be the
\code{Unknown} types}

\item{exprs}{a list of formulae with a predicate on the LHS and a PCV group name on
the RHS. which are interpreted as the parameters for a \code{dplyr::case_when}
call. This must be protected against interpretation by wrapping it in
\code{rlang::exprs()}. The predicates are tested against \code{avoncap::serotype_data$map}
and could use any of  the following columns 'serotype', 'PCV7', 'PCV13', 'PCV15', 'PCV20', 'PPV23', 'PCV13on7', 'PCV15on13', 'PCV20on15', 'PPV23on20', 'PCV10SSI', 'PCV10GSK', 'PCV15Zhifei', 'PCV24Vaxcyte', 'PCV24Affinivax'
a default option of the form \code{TRUE ~ "Non PCV serotype"} must exist to capture
unmatched items.}
}
\value{
A ggplot2 scale
}
\description{
The scale groups colours by PCV group, but it is important to have the source
data using the same levels as this scale otherwise the colour legend will be
ordered in a different sequence. This can be achieved using \code{relevel_serotypes},
}
