% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-location.R
\name{set_input}
\alias{set_input}
\title{Sets the location of data for an analysis}
\usage{
set_input(path)
}
\arguments{
\item{path}{the path to the input directory}
}
\value{
the full path to the directory
}
\description{
Also performs some structure checks and makes sure that the
README files are in place.
}
