% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked-bars.R
\name{stacked_barplot}
\alias{stacked_barplot}
\title{Stacked bar plot}
\usage{
stacked_barplot(data, mapping, ...)
}
\arguments{
\item{data}{the data}

\item{mapping}{a aes mapping with at least \code{x} and \code{fill}. If facetting then
\code{group} must contain the facet variable}

\item{...}{passed to \code{geom_bar}}
}
\value{
a ggplot
}
\description{
This function plots a stacked bar of proportions for an input set of data
}
\examples{
stacked_barplot(diamonds, aes(x=cut, fill=clarity, group=color))+facet_wrap(vars(color))

}
