% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-wrangling.R
\name{start_date_of_week}
\alias{start_date_of_week}
\title{Convert a study week back into a date}
\usage{
start_date_of_week(study_week)
}
\arguments{
\item{study_week}{does accept decimals and returns the nearest whole date to the value}
}
\value{
a vector of study_week numbers starting at zero for the first week of the study
}
\description{
This is poorly named as only give the start date is the input is an integer
}
