% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-wrangling.R
\name{study_week}
\alias{study_week}
\title{Convert a date to a study week}
\usage{
study_week(dates)
}
\arguments{
\item{dates}{a list of date objects}
}
\value{
an integer number of weeks since 2019-12-30
}
\description{
Convert a date to a study week
}
