% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-upset-plot.R
\name{upset_plot}
\alias{upset_plot}
\title{Upset plot with counts stratified by a categorical column}
\usage{
upset_plot(df, boolean_cols, categorical_col, lbl_size = 5)
}
\arguments{
\item{df}{the data}

\item{boolean_cols}{a tidyselect specification selecting the columns to be
used as binary one-hot encoded classes}

\item{categorical_col}{a column containing a disjoint category as a factor}

\item{lbl_size}{font sise of the label}
}
\value{
a ggplot
}
\description{
Upset plot with counts stratified by a categorical column
}
