% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-location.R
\name{valid_inputs}
\alias{valid_inputs}
\title{A valid set of types of file that can be loaded by \code{load_data(...)}}
\usage{
valid_inputs()
}
\value{
a dataframe of type, subtype
}
\description{
A valid set of types of file that can be loaded by \code{load_data(...)}
}
\examples{

# devtools::load_all()

avoncap::set_input("~/Data/avoncap")
avoncap::input("nhs-extract")

avoncap::all_files()


# exact match on filename column of all_data()
avoncap::most_recent_files("AvonCAPLRTDCentralDa")


# or matches by lower case startWith on directory
avoncap::most_recent_files("nhs-extract","deltave")


avoncap::most_recent_files("metadata")

avoncap::valid_inputs()
}
