% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework-validate.R
\name{validate_data}
\alias{validate_data}
\title{Validate AvonCap raw data}
\usage{
validate_data(rawData, ...)
}
\arguments{
\item{rawData}{\itemize{
\item the raw data from \code{load_data()}
}}

\item{...}{not used / passed to the validation function specific to
the type of data.}
}
\value{
the same input with a new \code{data_quality_failures} attribute containing
issues.
}
\description{
Runs a set of QA checks. This function dispatches the call
in a data set specific function using the \code{type} and \code{subtype} of the
data set. The checks are in source files named \code{validate-xxx.R} depending
on the data source.
}
