% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework-validate.R
\name{write_issues}
\alias{write_issues}
\title{Write out data quality issues}
\usage{
write_issues(df, file)
}
\arguments{
\item{df}{the raw data frame}

\item{file}{the output data quality file}
}
\value{
the list of failures as a dataframe
}
\description{
Write out data quality issues
}
