% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug-utils.R
\name{xglimpse}
\alias{xglimpse}
\title{Wrapper around \code{table}}
\usage{
xglimpse(data, ...)
}
\arguments{
\item{data}{a dataframe}

\item{...}{columns or named expressions to cross-tabulate}
}
\value{
the cross-tabulation
}
\description{
Wrapper around \code{table}
}
