% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoncap-augment.R
\name{augment_data}
\alias{augment_data}
\title{Augment dataset with derived data items}
\usage{
augment_data(avoncap_original)
}
\arguments{
\item{avoncap_original}{the data set as output from \code{avoncap::normalise_data()}}
}
\value{
a data set with extended derived fields
}
\description{
Augment dataset with derived data items
}
