% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_covid_status}
\alias{derive_covid_status}
\title{Determine if an admission is proven SARS-CoV-2 PCR positive}
\usage{
derive_covid_status(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
SARS-CoV-2 PCR positive only lab confirmed diagnosis.
}
\details{
Patient reported, clinical diagnoses are assumed PCR negative (although possible in
some cases they may not have been done).

In general cases where the test was not done will be NA.
}
\concept{derived}
