% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_covid_status}
\alias{derive_covid_status}
\title{Determine if an admission is proven SARS-CoV-2 PCR positive}
\usage{
derive_covid_status(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
SARS-CoV-2 PCR positive only lab confirmed diagnosis.
}
\details{
admission.covid_pcr_result:
\itemize{
\item based on fields: c19_adm_swab and covid_19_diagnosis
\item Patient reported, clinical diagnoses are assumed PCR negative (although possible in
some cases they may not have been done).
\item Lateral flows done in hospital are
counted as PCR negative.
\item negative admission swabs are counted as negative
\item NA signifies test not done.
}

admission.is_covid:
\itemize{
\item Binary confirmed or no-evidence.
\item PCR results count as confirmed,
\item Lateral flow results count as confirmed,
\item anything else is no evidence (includes negatives and test not done)
}
}
\concept{derived}
