% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_haematology_categories}
\alias{derive_haematology_categories}
\title{Binary outcomes for haematology data}
\usage{
derive_haematology_categories(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
\itemize{
\item Elevated troponin : > 18:  18ng/L is simply the 99th percentile value Beckman
assay we use as quoted by the IFCC. We elected to not use sex-specific 99th
percentile values although they are also quoted here and you could
incorporate into your analysis. I am sure you are aware of the 4th Universal
definition of MI that requires a rise or fall above the 99th percentile etc.
}
}
\concept{derived}
