% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-pneumococcal.R
\name{derive_phe_pcv_group}
\alias{derive_phe_pcv_group}
\title{Get vaccine coverage group for known serotype}
\usage{
derive_phe_pcv_group(
  df,
  ...,
  pcv_map = serotype_pcv_map,
  not_matched = "Other",
  col_name = "pneumo.phe_pcv_group"
)
}
\arguments{
\item{df}{the normalised longitudinal pneumococcal data}

\item{...}{ignored}

\item{pcv_map}{a 2 column data frame mapping \code{group} to \code{serotype}}

\item{not_matched}{what to call the column of non-matched serotypes? Default is
\code{Other}, but \verb{Non vaccine type} might be preferred.}

\item{col_name}{the target column name for the pcv grouping (defaults
to \code{pneumo.phe_pcv_group})}
}
\value{
an augmented data frame with an additional column defined by \code{col_name}
}
\description{
For the longitudinal oneumocococcal data, a range of useful serotype groups
is defined in the list \code{uad_groups}. The \code{default_pcv_map} gives a set of
mappings to group headings that gives the overall serotype distribution by
vaccine.
}
\concept{derived}
