% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_pneumococcal_risk_category}
\alias{derive_pneumococcal_risk_category}
\title{Determine pneumococcal risk group}
\usage{
derive_pneumococcal_risk_category(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
Original algorithm from B1851202 SAP defines a 3 class risk group
}
\details{
High-risk (immunocompromised)
\itemize{
\item ✗ Asplenia
\item ✓ Cancer/Malignancy, Hematologic
\item ✓ Cancer/Malignancy, Solid Tumor
\item ✓ Chronic Kidney Disease
\item ✓ Human Immunodeficiency Virus (HIV) – AIDS
\item ✓ Human Immunodeficiency Virus (HIV) – No AIDS
\item ✓ Immunodeficiency
\item ✓ Immunosuppressant Drug Therapy
\item ✓ Organ Transplantation
\item ✗ Multiple Myeloma
}

At Risk (immunocompetent)
\itemize{
\item ✓ Asthma
\item ✓ Alcoholism
\item ✗ Celiac Disease
\item ✓ Chronic Liver Disease without Hepatic Failure
\item ✓ Chronic Liver Disease with Hepatic Failure
\item ✓ Chronic Obstructive Pulmonary Disease
\item ✗ Cochlear Implant
\item ✓ Congestive Heart Failure
\item ✓ Coronary Artery Disease (CAD)
\item ✓ Chronic Neurologic Diseases
\item ✗ Coagulation factor replacement therapy
\item ✗ CSF Leak
\item ✓ Diabetes Treated with Medication
\item ✓ Down syndrome
\item ✓ Institutionalized in nursing home or LTC facility (Nursing home or long-term care
facility for those with disability or dependency on subject characteristics/risk
determinants eCRF page)
\item ✓ Occupational risk with exposure to metal fumes
\item ✓ Other Chronic Heart Disease
\item ✓ Other Chronic Lung Disease
\item ✓ Other pneumococcal disease risk factors
\item ✗ Previous Invasive Pneumococcal Disease
\item ✓ Tobacco smoking (Tobacco/E-Cigarettes)
}

Anything else is low risk
}
\concept{derived}
