% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_severe_disease_outcomes}
\alias{derive_severe_disease_outcomes}
\title{Binary outcomes for severe disease}
\usage{
derive_severe_disease_outcomes(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
\itemize{
\item Confirmed death (any length follow up)
\item Any ICU admission
}
}
\details{
described in aLRTD paper
}
\concept{derived}
