% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-avoncap.R
\name{derive_simpler_comorbidities}
\alias{derive_simpler_comorbidities}
\title{Rationalise some of the more detailed comorbidities}
\usage{
derive_simpler_comorbidities(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
and generate some summary values
}
\details{
\itemize{
\item simple DM without insulin dependence
\item Solid / Haematological / Any cancer present binary indicators
\item any chronic resp dx: i.e. any of asthma, bronchiectasis, chronic pleural disease, COPD,
interstitial lung dx, cyctic fibrosis, other chronic resp dx
\item any chronic heart disease: pulmonary htn, CCF, IHD, previous MI, congential
heart dx, hypertension, AF, other arrythmia, other heart dx, other other heart dx
\item Stroke or TIA binary
\item Any immune compromise binary (immunodeficient or on immune suppressants)
}
}
\concept{derived}
