% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-generic.R
\name{derive_survival_censoring}
\alias{derive_survival_censoring}
\title{Survival outcomes}
\usage{
derive_survival_censoring(df, v, ...)
}
\arguments{
\item{df}{the dataframe.}

\item{v}{the value set. usually precomputed by the augment framework the value
set can be explicitly supplied with \code{v = get_value_sets(df)}}

\item{...}{ignored}
}
\value{
a dataframe
}
\description{
Expects as days since admission:
\itemize{
\item \code{survival.length_of_stay} - length of stay until discharge or death (NA if still in hosptial),
\item \code{survival.uncensored_time_to_death} - time until death (NA if alive at last obs),
\item \code{survival.last_observed_event} - last time patient observed alive.
}
}
\details{
Calculates
\itemize{
\item a 30 day survival duration and censoring status for survfit
\item a 1 year survival duration and censoring status for survfit
\item Hospital length of stay and censoring status for survfit
\item Categorical length of stay and 30 day survival 0-3, 4-6, 7-13, 14-29, gte 30
}

Survival data will be of the form:

\code{survival.30_day_death_xxx}, \code{survival.1_yr_death_xxx}, \code{survival.30_day_discharge}

\code{xxx_time}: for this is the follow up time to event in days (max 30 or 365).

\code{xxx_event}: The event type indicator
\itemize{
\item 0 = alive at event (censored),
\item 1 = dead.
}

or for length of stay:
\itemize{
\item 0 = still inpatient / died (censored),
\item 1 = discharged from hospital
}

A survival model will be of the form:

\code{survival::Surv(time = xxx_time, event=xxx_event) ~ ...}
}
\concept{derived}
