% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-labelling.R
\name{extract_dependencies}
\alias{extract_dependencies}
\title{Extract the}
\usage{
extract_dependencies(data, col, original = TRUE)
}
\arguments{
\item{data}{the dataframe}

\item{col}{the column as a symbol}

\item{original}{map the names to the original column names from the data. If
this is false the function returns a list of current normalised column names.}
}
\value{
a named list of dependencies and original column names for a given
column
}
\description{
Extract the
}
