% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km-plot.R
\name{km_plot}
\alias{km_plot}
\title{Faceted Kaplan-Meier plot}
\usage{
km_plot(
  df,
  coxmodel,
  facet = NULL,
  ...,
  maxtime = max(sfit$time),
  ylab = if (!invert) "surviving (\%)" else "affected (\%)",
  xlab = "time (days)",
  ylim = c(NA, 100),
  n_breaks = 5,
  heights = c(10, 1),
  invert = FALSE
)
}
\arguments{
\item{df}{the data}

\item{coxmodel}{the cox model output of survival::coxph from the data}

\item{facet}{the division to highlight in the KM strata. Defaults to first
term on the lhs of the cox model formula}

\item{...}{
  Arguments passed on to \code{\link[survival:survfit]{survival::survfit}}
  \describe{
    \item{\code{formula}}{either a formula or a previously fitted model}
  }}

\item{maxtime}{the longest x value to plot}

\item{ylab}{the y axis label}

\item{xlab}{the x axis label}

\item{ylim}{the range to show on the KM plot}

\item{n_breaks}{number of x axis breaks to display this also determines the
timing and number of "at risk" counts to display.}

\item{heights}{the relative height between the KM plot and the "at risk" table}

\item{invert}{reverse survival statistics to count number of affected}
}
\value{
a ggplot patchwork.
}
\description{
Faceted Kaplan-Meier plot
}
