% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wrangling.R
\name{load_data}
\alias{load_data}
\title{Load data and check structure}
\usage{
load_data(
  type,
  file = NULL,
  reproduce_at = as.Date(getOption("reproduce.at", default = Sys.Date())),
  merge = NA,
  ...
)
}
\arguments{
\item{type}{see valid_inputs() for current list in input directory}

\item{file}{see valid_inputs() for current list in input directory}

\item{reproduce_at}{\itemize{
\item the date at which to cut off newer data files
}}

\item{merge}{\itemize{
\item setting to \code{TRUE} forces multiple files be merged into a single data frame by losing mismatching columns.
}}

\item{...}{\itemize{
\item passed to \code{cached} may specifically want to use `nocache=TRUE``
}}
}
\value{
either a list of dataframes or a single merged dataframe
}
\description{
Loads the AvonCap data from a set of csv files, which may optionally be qualified by site \verb{('BRI' or 'NBT')} and
database year \verb{('20-21','21-22','22-23', or 'y1', 'y2', 'y3')} as part of the file name. This selects the most
recent files earlier than the \code{reproduce_at} date and detects whether they are in a set of files.
}
\details{
The files are loaded as csv as checked that files have (A) the same columns, (B) the same type (or are empty) (C)
have any major parse issues. It then merges the files into a single dataframe, if possible, otherwise it will
return the individually loaded files as a list of dataframes.
}
\examples{
load_data("nhs-extract")
}
