% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-avoncap-mappings.R
\name{map_avoncap_central}
\alias{map_avoncap_central}
\title{Core avoncap normalisation}
\usage{
map_avoncap_central()
}
\value{
a list
}
\description{
\itemize{
\item record_number -> admin.record_number (name)
\item what_was_the_first_surveil -> admin.first_record_number (name)
\item ac_study_number -> admin.consented_record_number (study_id)
\item nhs_number -> admin.patient_identifier (ppi)
\item duplicate -> admin.duplicate (yesno)
\item enrollment_date -> admin.enrollment_date (date)
\item study_year -> admin.study_year (name)
\item file -> admin.data_file (name)
\item week_number -> admin.week_number (name)
\item c19_diagnosis -> diagnosis.standard_of_care_COVID_diagnosis (list)
\item clinical_radio_diagnosis -> diagnosis.clinical_or_radiological_LRTI_or_pneumonia (yesno)
\item c19_adm_swab -> diagnosis.admission_swab (list)
\item c19_test_type -> diagnosis.test_type (list)
\item qualifying_symptoms_signs -> diagnosis.qualifying_symptoms_signs (name)
\item cc_critieria -> diagnosis.meets_case_control_criteria (yesno)
\item cc_pos_date -> diagnosis.first_COVID_positive_swab_date (date)
\item gender -> demog.gender (list)
\item age_at_admission -> demog.age (double)
\item age_march -> demog.age_in_march_2021 (double)
\item imd -> demog.imd_decile (name)
\item smoking -> demog.smoker (list)
\item ethnicity2 -> demog.ethnicity (list)
\item care_home -> demog.care_home_resident (yesno)
\item drugs -> dplyr (checkboxes)
\item vaping -> demog.vaping (list)
\item alc_units -> demog.units_of_alcohol (name)
\item np_swab -> admin.np_swab_taken_1 (list)
\item adm_np_type -> admin.np_swab_site_1 (list)
\item np_date -> admin.np_swab_date_1 (date)
\item days_adm_npswab -> admin.np_swab_day_since_admission (double)
\item np_swab_2 -> admin.np_swab_taken_2 (list)
\item adm_np_type_2 -> admin.np_swab_site_2 (list)
\item np_date_2 -> admin.np_swab_date_2 (date)
\item np_swab_3 -> admin.np_swab_taken_3 (list)
\item adm_np_type_3 -> admin.np_swab_site_3 (list)
\item np_date_3 -> admin.np_swab_date_3 (date)
\item saliva -> admin.saliva_sample_taken (list)
\item saliva_date -> admin.saliva_sample_date (date)
\item days_adm_saliva -> admin.saliva_sample_day_since_admission (double)
\item sputum -> admin.sputum_sample_taken (list)
\item sputum_date -> admin.sputum_sample_date (date)
\item days_adm_sputum -> admin.sputum_sample_day_since_admission (double)
\item pt_ad_ur -> admin.urine_sample_needed (yesno)
\item adm_ur_taken -> admin.urine_sample_taken (list)
\item nourine_reason -> admin.urine_sample_failure_reason (list)
\item adm_np_type_2 -> admin.urine_sample_site (list)
\item adm_ur_date -> admin.urine_sample_date (date)
\item days_adm_urine -> admin.urine_sample_day_since_admission (double)
\item adm_serum_tak -> admin.serum_sample_taken (list)
\item adm_seru_date -> admin.serum_sample_date (date)
\item days_adm_serum -> admin.serum_sample_day_since_admission (double)
\item contraindication -> vaccination.covid_vaccine_contraindicated (yesno)
\item covid19_vax -> vaccination.covid_vaccination (list)
\item covidvax_date -> vaccination.first_dose_date (date)
\item covidvax_dose_2 -> vaccination.second_dose_date (date)
\item covidvax_dose_3 -> vaccination.third_dose_date (date)
\item covidvax_dose_4 -> vaccination.fourth_dose_date (date)
\item covidvax_dose_5 -> vaccination.fifth_dose_date (date)
\item covidvax_dose_6 -> vaccination.sixth_dose_date (date)
\item brand_of_covid19_vaccinati -> vaccination.first_dose_brand (list)
\item covid19vax_brand_2 -> vaccination.second_dose_brand (list)
\item covid19vax_brand_3 -> vaccination.third_dose_brand (list)
\item covid19vax_brand_4 -> vaccination.fourth_dose_brand (list)
\item covid19vax_brand_5 -> vaccination.fifth_dose_brand (list)
\item covid19vax_brand_6 -> vaccination.sixth_dose_brand (list)
\item c19vaxd1_adm -> admission.time_since_first_vaccine_dose (name)
\item c19vaxd2_adm -> admission.time_since_second_vaccine_dose (name)
\item c19vaxd3_adm -> admission.time_since_third_vaccine_dose (name)
\item c19vaxd4_adm -> admission.time_since_fourth_vaccine_dose (name)
\item c19vax5_adm -> admission.time_since_fifth_vaccine_dose (name)
\item c19vax6_adm -> admission.time_since_sixth_vaccine_dose (name)
\item flu_date -> vaccination.last_flu_dose_date (date)
\item fluvax_adm_d1 -> admission.time_since_last_flu_vaccine_dose (name)
\item ppv23_date -> vaccination.last_pneumococcal_dose_date (date)
\item ppv23vax_adm_d -> admission.time_since_last_pneumococcal_vaccine_dose (name)
\item c19_variant -> genomic.variant (variant)
\item year -> admission.year (double)
\item study_week -> admission.study_week (double)
\item admission_date -> admission.date (date)
\item hospital -> admin.hospital (text_to_factor)
\item ics -> admission.on_inhaled_corticosteroids (yesno)
\item immsup -> admission.on_immunosuppression (yesno)
\item psi_class -> admission.pneumonia_severity_index_class (list)
\item crb_test_mai -> admission.curb_65_severity_score (list)
\item news_2_total -> admission.news2_score (name)
\item pulse_ox -> admission.oximetry (name)
\item rr -> admission.respiratory_rate (name)
\item fio2 -> admission.max_oxygen (name)
\item systolic_bp -> admission.systolic_bp (name)
\item diastolic_bp -> admission.diastolic_bp (name)
\item hr -> admission.heart_rate (name)
\item temperature -> admission.temperature (list)
\item symptom_days_preadmit -> admission.duration_symptoms (name)
\item previous_infection -> admission.previous_covid_infection (list)
\item previousinfection_date -> admission.previous_covid_infection_date (date)
\item c19d_preadm -> admission.time_since_covid_diagnosis (name)
\item rockwood -> admission.rockwood_score (name)
\item cci_total_score -> admission.charlson_comorbidity_index (name)
\item height -> admission.height (name)
\item weight -> admission.weight (name)
\item bmi -> admission.BMI (double)
\item first_radio -> dplyr (checkboxes)
\item c19_peep -> day_7.max_peep (name)
\item c19_hospadm -> day_7.length_of_stay (list)
\item c17_high -> day_7.max_care_level (list)
\item c19icuon -> day_7.still_on_icu (yesno)
\item c19_icudays -> day_7.icu_length_of_stay (list)
\item c19_vent -> day_7.max_ventilation_level (list)
\item c19_ox -> day_7.max_o2_level (list)
\item c19_ionotropes -> day_7.ionotropes_needed (yesno)
\item c19_complication -> dplyr (checkboxes)
\item c19_death7d -> day_7.death (yesno)
\item c19_meds -> dplyr (checkboxes)
\item hospital_length_of_stay -> outcome.length_of_stay (name)
\item survival_days -> outcome.survival_duration (name)
\item ip_death -> outcome.inpatient_death (yesno)
\item days_in_icu -> outcome.icu_duration (name)
\item did_the_patient_have_respi -> outcome.respiratory_support_needed (yesno)
\item number_of_days_of_ventilat -> outcome.ventilator_duration (name)
\item ett_days -> outcome.endotracheal_tube_duration (name)
\item renal_replacement_therapy -> outcome.renal_support_duration (name)
\item complications -> dplyr (checkboxes)
\item ventilatory_support -> outcome.highest_level_ventilatory_support (list)
\item did_the_patient_receive_ec -> outcome.received_ecmo (yesno)
\item inotropic_support_required -> outcome.received_ionotropes (yesno)
\item lrtd_30d_outcome -> outcome.functional_status (list)
\item fever2 -> symptom.abnormal_temperature (yesno_unknown)
\item pleurtic_cp -> symptom.pleuritic_chest_pain (yesno_unknown)
\item cough2 -> symptom.cough (yesno_unknown)
\item sput_prod -> symptom.productive_sputum (yesno_unknown)
\item dyspnoea -> symptom.dyspnoea (yesno_unknown)
\item tachypnoea2 -> symptom.tachypnoea (yesno_unknown)
\item anosmia -> symptom.anosmia (yesno_unknown)
\item ageusia -> symptom.ageusia (yesno_unknown)
\item dysgeusia -> symptom.dysguesia (yesno_unknown)
\item fever -> symptom.fever (yesno_unknown)
\item hypothermia -> symptom.hypothermia (yesno_unknown)
\item chills -> symptom.chills (yesno_unknown)
\item headache -> symptom.headache (yesno_unknown)
\item malaise -> symptom.malaise (yesno_unknown)
\item wheeze -> symptom.wheeze (yesno_unknown)
\item myalgia -> symptom.myalgia (yesno_unknown)
\item worse_confusion -> symptom.worsening_confusion (yesno_unknown)
\item general_det -> symptom.general_deterioration (yesno_unknown)
\item ox_on_admission -> symptom.oxygen_required_on_admission (yesno_unknown)
\item resp_disease -> dplyr (checkboxes)
\item other_respiratory_disease -> dplyr (checkboxes)
\item chd -> dplyr (checkboxes)
\item mi -> comorbid.previous_mi (yesno)
\item other_chd -> dplyr (checkboxes)
\item diabetes -> comorbid.diabetes (list)
\item dm_meds -> comorbid.diabetes_medications (list)
\item neurological_disease -> dplyr (checkboxes)
\item dementia -> dplyr (checkboxes)
\item cancer -> comorbid.solid_cancer (list)
\item haem_malig -> dplyr (checkboxes)
\item ckd -> comorbid.ckd (list)
\item liver_disease -> comorbid.liver_disease (list)
\item gastric_ulcers -> comorbid.gastric_ulcers (yesno)
\item pvd -> comorbid.periph_vasc_dx (yesno)
\item ctd -> comorbid.connective_tissue_dx (yesno)
\item immunodeficiency -> comorbid.immunodeficiency (yesno)
\item other_pn_disease -> comorbid.other_pneumococcal_risks (yesno)
\item transplant -> comorbid.transplant_recipient (yesno)
\item pregnancy -> comorbid.pregnancy (list)
\item hiv -> dplyr (checkboxes)
\item final_soc_lrtd_diagnosis -> renameToVars (checkboxes)
\item covid_19_diagnosis -> diagnosis.covid_19_diagnosis (list)
\item ppv23 -> vaccination.pneumovax (list)
\item flu_vaccine -> vaccination.influenza_vaccination (list)
}
}
\concept{map}
