% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-pneumo-mappings.R
\name{map_avoncap_pneumococcal}
\alias{map_avoncap_pneumococcal}
\title{Normalise the avoncap pneumococcal data}
\usage{
map_avoncap_pneumococcal()
}
\value{
a list
}
\description{
\itemize{
\item participant_number -> admin.record_number (name)
\item hospital -> admin.hospital (list)
\item nhs_number -> admin.patient_identifier (ppi)
\item age_at_admission -> demog.age (double)
\item sex -> demog.gender (list)
\item test_date -> pneumo.test_date (date)
\item test -> pneumo.test_type (list)
\item serotype -> pneumo.phe_serotype (pneumo_serotype)
\item smoker -> demog.smoker (list)
\item resp_disease -> dplyr (checkboxes)
\item chd -> dplyr (checkboxes)
\item mi -> comorbid.previous_mi (yesno)
\item ckd -> comorbid.ckd (list)
\item liver_disease -> comorbid.liver_disease (list)
\item diabetes -> comorbid.diabetes (list)
\item dm_meds -> comorbid.diabetes_medications (list)
\item dementia -> dplyr (checkboxes)
\item neurological_disease -> dplyr (checkboxes)
\item gastric_ulcers -> comorbid.gastric_ulcers (yesno)
\item dysphagia -> comorbid.dysphagia (yesno)
\item pvd -> comorbid.periph_vasc_dx (yesno)
\item ctd -> comorbid.connective_tissue_dx (yesno)
\item immunodeficiency -> comorbid.immunodeficiency (yesno)
\item other_pn_disease -> comorbid.other_pneumococcal_risks (yesno)
\item hiv -> dplyr (checkboxes)
\item cancer -> comorbid.solid_cancer (list)
\item haem_malig -> dplyr (checkboxes)
\item recent_chemo -> comorbid.recent_chemotherapy (yesno)
\item recent_radiotherapy -> comorbid.recent_radiotherapy (yesno)
\item transplant -> comorbid.transplant_recipient (yesno)
\item pregnancy -> comorbid.pregnancy (list)
\item drugs -> dplyr (checkboxes)
\item immsup -> admission.on_immunosuppression (yesno)
\item weight_problem -> comorbid.bmi_status (list)
\item concomittant_flu -> comorbid.influenza_infection (yesno)
\item hcv -> comorbid.hepatitis_c (yesno)
\item ppv23 -> vaccination.ppv (list)
\item flu_vaccine -> vaccination.flu (list)
\item cci_total_score -> admission.charlson_comorbidity_index (name)
\item los_days -> outcome.length_of_stay (double)
\item amts -> admission.triage_score (list)
\item resp_rate -> admission.respiratory_rate (double)
\item sats_ra -> admission.saturations_on_room_air (double)
\item systolic_bp -> admission.systolic_bp (double)
\item diastolic_bp -> admission.diastolic_bp (double)
\item crb65_score -> admission.crb65 (double)
\item curb65_score -> admission.curb65 (double)
\item antibiotic_route -> outcome.antibiotic_route (list)
\item antibiotic_days -> outcome.antibiotic_duration (double)
\item infection_site -> admission.infection_site (list)
\item deranged_lfts -> outcome.abnormal_lft (yesno)
\item aki -> outcome.akute_kidney_injury (yesno)
\item pleural_effusion -> outcome.pleural_effusion (yesno)
\item empyema -> outcome.empyema (yesno)
\item discharge_destination -> outcome.discharge_to (list)
\item icu -> outcome.admitted_icu (yesno)
\item niv -> outcome.non_invasive_ventilation (yesno)
\item intubation -> outcome.intubation (yesno)
\item recurrent_pneumonia -> outcome.recurrent_pneumonia (yesno)
\item ecmo -> outcome.received_ecmo (yesno)
\item inotropes -> outcome.received_ionotropes (yesno)
\item trachy -> outcome.tracheostomy (yesno)
\item inpatient_death -> outcome.inpatient_death (yesno)
\item death_30days -> outcome.death_within_30_days (yesno)
\item death_1year -> outcome.death_within_1_year (yesno)
\item survival_days -> outcome.survival_duration (name)
\item albumin -> haem.albumin (double)
\item wcc -> haem.white_cell_count (double)
\item hb -> haem.haemoglobin (double)
\item pmn -> haem.neutrophils (double)
\item lymphocytes -> haem.lymphocytes (double)
\item crp -> haem.crp (double)
\item na_result -> haem.sodium (double)
\item ur_result -> haem.urea (double)
\item egfr -> haem.egfr (double)
\item creatinine -> haem.creatinine (double)
\item cxr_sides -> radio.cxr_infection (list)
\item cxr_lobes -> radio.cxr_lobar_changes (list)
\item death_5year -> outcome.death_within_5_years (yesno)
\item survival_days_2 -> outcome.5_yr_survival_duration (name)
\item imd_decile -> demog.imd_decile (name)
}
}
\concept{map}
