% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-pneumo-mappings.R
\name{map_avoncap_pneumococcal}
\alias{map_avoncap_pneumococcal}
\title{Normalise the avoncap pneumococcal data}
\usage{
map_avoncap_pneumococcal()
}
\value{
a list
}
\description{
\itemize{
\item test_date -> pneumo.test_date (date)
\item test -> pneumo.test_type (list)
\item serotype -> pneumo.phe_serotype (pneumo_serotype)
}
}
\concept{map}
