% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-avoncap-virol-mappings.R
\name{map_avoncap_virol}
\alias{map_avoncap_virol}
\title{Normalise the avoncap data virology data}
\usage{
map_avoncap_virol(instrument)
}
\arguments{
\item{instrument}{the numeric instrument number}
}
\value{
a list
}
\description{
\itemize{
\item viral_testing_performed -> virol.test_performed (yesno)
\item virology_date_of_asst -> virol.test_date (date)
\item viroldays -> virol.test_days_from_admission (pos_integer)
\item specimen_type -> virol.test_type (list)
\item virus_isolated -> virol.pathogen_detected (yesno)
\item test_type -> virol.test_type (list)
\item virus_pathogen -> virol.pathogen, .virol_isolate_list (checkboxes_to_nested_list)
\item virol_patient_lab -> virol.test_provenance (list)
}
}
\concept{map}
