% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping-uad-binax.R, R/mapping-uad-serotypes.R
\name{map_urine_binax}
\alias{map_urine_binax}
\title{Normalise the urinary antigen data (binax results)}
\usage{
map_urine_binax()

map_urine_binax()
}
\value{
a list

a list
}
\description{
\itemize{
\item RESULT -> pneumo.binax_result, .x (text)
\item EVENT_DATE -> pneumo.test_date (date)
\item SUBJECT -> admin.consented_record_number (study_id)
\item BARCODE -> pneumo.urine_antigen_sample_id (name)
}

\itemize{
\item RESULT -> pneumo.binax_result, .x (text)
\item EVENT_DATE -> pneumo.test_date (date)
\item SUBJECT -> admin.consented_record_number (study_id)
\item BARCODE -> pneumo.urine_antigen_sample_id (name)
}
}
\concept{map}
