% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-location.R
\name{most_recent_files}
\alias{most_recent_files}
\title{find most recent files of a specific type}
\usage{
most_recent_files(
  type = "",
  file = NULL,
  reproduce_at = as.Date(getOption("reproduce.at", default = Sys.Date()))
)
}
\arguments{
\item{type}{see valid_inputs() for current list of supported types in input directory}

\item{file}{see valid_inputs() for list of supported filenames}

\item{reproduce_at}{after this date new files are ignored. This enforces a specific version of the data.}
}
\value{
a list of the file paths to the most up to date files of the given type relevant to each site and study year
}
\description{
find most recent files of a specific type
}
\examples{

# devtools::load_all()

avoncap::set_input("~/Data/avoncap")
avoncap::input("nhs-extract")

avoncap::all_files()


# exact match on filename column of all_data()
avoncap::most_recent_files("AvonCAPLRTDCentralDa")


# or matches by lower case startWith on directory
avoncap::most_recent_files("nhs-extract")


avoncap::most_recent_files("metadata")

avoncap::valid_inputs()
}
