% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mappers.R
\name{normalise_data}
\alias{normalise_data}
\title{Sanitise AvonCap data columns}
\usage{
normalise_data(
  rawData,
  ethn = try(avoncap::load_data("ethnicity", reproduce_at = Sys.Date()), silent = TRUE),
  remove_mapped = TRUE,
  remove_unmapped = TRUE,
  mappings = .mappings,
  messages = c("files: {files}", "{.total} rows", "analysis date: {reproduce_at}",
    "files from: {date}"),
  reproduce_at = as.Date(getOption("reproduce.at", default = Sys.Date())),
  data_source_info = NULL,
  ...
)
}
\arguments{
\item{rawData}{\itemize{
\item the raw data from \code{load_data()}
}}

\item{ethn}{the supplementary ethnicity files. but default will use the call \code{avoncap::most_recent_files("ethnicity")} to find these files and will- can be set to "NULL" if ethnicity files are not available.}

\item{remove_mapped}{gets rid of original columns for which we have a mapping (leaving the new versions)}

\item{remove_unmapped}{gets rid of columns for which we do not have a mapping}

\item{mappings}{a set of mappings (see \code{zzz-avoncap-mappings.R})}

\item{messages}{a set of \code{dtrackr} glue specs that populate the first box fo the flow chart. (can use \code{{files}}, \code{{reproduce_at}}, \code{{date}}, \code{{.total}})}

\item{reproduce_at}{the date at which to reproduce the analysis to (\code{options(reproduce.at = "...")})}

\item{data_source_info}{\itemize{
\item if not null a filename, and the function will write out a file with the details of the input files used.
}}

\item{...}{passed onto \code{.cached(...)}. e.g. \code{nocache = TRUE} can be used to defeat caching.}
}
\value{
a tracked dataframe with
}
\description{
AvonCap data has lots of columns which are named in a difficult to remember fashion, composed of
data items that have enumerated values with no semantics. This makes displaying them difficult and
any filtering done on the raw data inscrutable. Depending on the source of the data some different columns
may be present due to differences in the NHS and UoB data sets. The redcap database has some options that may be
checklists and some that are radio buttons, both of these end up with mysterious names in the data.
}
\details{
This function maps the data into a tidy dataframe with consistently named columns, and named factors where appropriate.
If not present in the data the ethnicity files
Most of the sanitisation code is held in the \code{zzz-avoncap-mappings.R} file.
}
