% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wrangling.R
\name{original_field_names}
\alias{original_field_names}
\title{Get the mapping of transformed columns back to original}
\usage{
original_field_names(data)
}
\arguments{
\item{data}{the transformed data set.}
}
\value{
a named list mapping original to new columns
}
\description{
Get the mapping of transformed columns back to original
}
